Update ID-RR
Reworked some of the logic. Your original code was including andi, ori, among a few other instructions that we do not need for this project. Changed all of the opcode comparisons to be directly with a binary number. Added an isValid variable to check if the opcode is valid, since your original code only checked if the final if statement succeeded, which only checks a few of the valid instructions. Finally, fixed the read register function, since you were not dereferencing the data pointers. 
One last note, I highly recommend that you restructure your decode function into one long if-else or switch statement to check each individual opcode. I didn't do that, in case you don't want to do that, but the way you have it now will make it harder to debug later on, since the control assignments are spread over 100 lines rather than localized within a single if-else statement or switch case for an individual opcode. It would also reduce redundant comparisons and (imo) be significantly easier to read.